requirements.txt
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
decorator==5.2.1
Flask==3.1.0
Flask-Caching==2.3.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
moviepy==2.1.2
numpy==2.2.4
passlib==1.7.4
pillow==10.4.0
proglog==0.1.11
python-dotenv==1.1.0
qrcode==8.1
schedule==1.2.2
SQLAlchemy==2.0.40
tinydb==4.8.2
tqdm==4.67.1
typing_extensions==4.13.2
validators==0.34.0
Werkzeug==3.1.3
yt-dlp==2025.3.31
